requirements.txt
django
djangorestframework==3.15.2
psycopg2-binary
celery[redis]
djoser
djangorestframework-simplejwt
Pillow
drf-yasg
django_celery_beat
django_celery_results
django-environ